refactor: modularize verification code
